requirements.txt
alembic==1.11.1
anyio==3.6.2
argon2-cffi==21.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bcrypt==4.0.1
bleach==4.1.0
blinker==1.6.2
caer==2.0.8
certifi==2022.12.7
cffi==1.14.6
charset-normalizer==2.0.4
click==8.1.3
colorama==0.4.4
cycler==0.11.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
entrypoints==0.3
et-xmlfile==1.1.0
exceptiongroup==1.1.0
face-recognition-models==0.3.0
fastapi==0.95.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
fonttools==4.28.5
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.2
importlib-metadata==6.6.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.0
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyter-server==1.11.0
jupyterlab==3.1.11
jupyterlab-pygments==0.1.2
jupyterlab-server==2.8.1
kiwisolver==1.3.2
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy==0.950
mypy-extensions==0.4.3
nbclassic==0.3.1
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.7.1
notebook==6.4.3
numpy==1.21.3
opencv-contrib-python==4.5.5.64
openpyxl==3.0.9
orjson==3.8.12
outcome==1.2.0
packaging==21.0
pandas==1.3.5
pandocfilters==1.4.3
parso==0.8.2
passlib==1.7.4
pickleshare==0.7.5
Pillow==9.0.0
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psycopg2==2.9.6
pyasn1==0.5.0
pycparser==2.20
pydantic==1.10.7
Pygments==2.10.0
pyodbc==4.0.34
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2021.1
pywin32==301
pywinpty==1.1.4
PyYAML==6.0
pyzmq==22.2.1
requests==2.26.0
requests-unixsocket==0.2.0
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
selenium==4.7.2
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.13
starlette==0.26.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
tomli==2.0.1
tornado==6.1
traitlets==5.1.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
ujson==5.7.0
urllib3==1.26.6
uvicorn==0.22.0
watchfiles==0.19.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.1
websockets==11.0.3
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0
